refactor backend structure
