+++ backend/services/__init__.py
@@ -1,7 +1,9 @@
 from .documents import DocumentService
 from .lines import LineService
+from .ocr import OcrService
 
 __all__ = [
     "DocumentService",
     "LineService",
+    "OcrService",
 ]
